test: update use-session-polling tests for long polling implementation
- Replace old polling tests with long polling version tests
- Add tests for version tracking and 204 No Content handling
- Test turn merging logic and error recovery scenarios
- Remove duplicate long-polling.test.tsx file
